threshold utility
Former-commit-id: 326467b2db8e1329ccf3db0110f65dddcb5fd5ce

+++ trunk/GeoImageViewer/src/main/java/org/geoimage/viewer/util/Utils.java
@@ -201,7 +201,7 @@ public class Utils {
      * @param thresholdArrayValues
      * @return
      */
-    public static String[] getStringThresholdsArray(SarImageReader reader,int[] thresholdArrayValues){
+    public static String[] getStringThresholdsArray(SarImageReader reader,float[] thresholdArrayValues){
     	int numberOfBands=reader.getNBand();
 	    final String[] thresholds = new String[numberOfBands];
 	    //management of the strings added at the end of the layer name in order to remember the used threshold
